training/pytorch/Palma_model.py
"""Legacy import path: re-exports the ConvLSTM stack from pytorch_convLSTM."""

from pytorch_convLSTM import HDF5SequenceDataset, MultiLayerConvLSTM

__all__ = ["HDF5SequenceDataset", "MultiLayerConvLSTM"]
